# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NEO APP API",
    "version": "1.0.0",
    "description": "REST API ของร้าน NEO APP (ขายแอปพลิเคชัน + กระเป๋าเงิน)\n\n## การยืนยันตัวตน\n\nทุก endpoint ต้องผ่านการยืนยันตัวตน ไม่มีเส้นไหนเปิดสาธารณะ รองรับ 2 วิธี:\n\n1. **Cookie session (เบราว์เซอร์)** — ล็อกอินที่ `/login` แล้ว Supabase จะเก็บ JWT ไว้ใน cookie `sb-<project-ref>-auth-token`\n2. **Bearer token (Postman / curl / script)** — ส่งเฮดเดอร์ `Authorization: Bearer <access_token>`\n\nขอ access token:\n\n```bash\ncurl -X POST \"$NEXT_PUBLIC_SUPABASE_URL/auth/v1/token?grant_type=password\" \\\n  -H \"apikey: $NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY\" \\\n  -H \"content-type: application/json\" \\\n  -d '{\"email\":\"you@example.com\",\"password\":\"...\"}'\n```\n\n## สิทธิ์\n\nข้อมูลถูกคุมด้วย Row Level Security ของ Postgres: ลูกค้าเห็นเฉพาะคำสั่งซื้อ/กระเป๋าเงินของตัวเอง\nส่วนการเพิ่ม-แก้-ลบสินค้า อัปโหลดรูป และแก้ตั้งค่าร้าน ทำได้เฉพาะบัญชีที่มี `app_metadata.role = \"admin\"`\n\n## เอกสารนี้\n\nหน้า `/docs` และไฟล์ `/openapi.json` เปิดให้เฉพาะแอดมิน คนอื่นจะได้ 404",
    "contact": {
      "name": "NEO APP"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "รายการสินค้า (กรอง/ค้นหา/จำกัดจำนวนได้)",
        "operationId": "listProducts",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "ชื่อหมวดหมู่ ใช้ค่า `ทั้งหมด` เพื่อไม่กรอง",
            "schema": {
              "type": "string"
            },
            "example": "เพิ่มประสิทธิภาพ"
          },
          {
            "name": "search",
            "in": "query",
            "description": "ค้นหาจากชื่อ หมวดหมู่ หรือคำอธิบาย",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "featured",
            "in": "query",
            "description": "ใส่ `true` เพื่อเอาเฉพาะสินค้าแนะนำหรือป้าย HOT",
            "schema": {
              "type": "string",
              "enum": [
                "true"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "จำกัดจำนวนรายการที่ส่งกลับ",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "example": 4
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "เพิ่มสินค้าใหม่",
        "operationId": "createProduct",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              },
              "example": {
                "name": "Focus Timer Pro",
                "category": "เพิ่มประสิทธิภาพ",
                "price": 299,
                "originalPrice": 499,
                "stock": 100,
                "description": "แอปจับเวลาโฟกัสแบบ Pomodoro",
                "specs": {
                  "รองรับ": "iOS, Android",
                  "อัปเดต": "ตลอดอายุการใช้งาน"
                },
                "badge": "NEW",
                "isFeatured": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "เพิ่มสินค้าสำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "อ่านการตั้งค่าร้าน",
        "operationId": "getSettings",
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/StoreSettings"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "แก้การตั้งค่าร้าน (แอดมินเท่านั้น)",
        "operationId": "updateSettings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/StoreSettings"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "ไม่ใช่แอดมิน",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/supplier": {
      "get": {
        "tags": [
          "Supplier"
        ],
        "summary": "แคตตาล็อกซัพพลายเออร์ + เครดิตคงเหลือ (แอดมิน)",
        "operationId": "listSupplierCatalogue",
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "account": {
                          "type": "object"
                        },
                        "products": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "ไม่ใช่แอดมิน"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "description": "ซัพพลายเออร์ตอบผิดพลาด"
          }
        }
      },
      "post": {
        "tags": [
          "Supplier"
        ],
        "summary": "นำเข้าสินค้าจากซัพพลายเออร์เข้าร้าน (แอดมิน)",
        "operationId": "importSupplierProduct",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": {
                    "type": "string"
                  },
                  "productIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "price": {
                    "type": "number",
                    "description": "ราคาขาย ถ้าไม่ระบุใช้ราคาแนะนำ"
                  },
                  "isFeatured": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "นำเข้าสำเร็จ"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "ไม่ใช่แอดมิน"
          },
          "404": {
            "description": "ไม่พบสินค้าในแคตตาล็อก"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Cookie session ของ Supabase (ชื่อจริงคือ `sb-<project-ref>-auth-token`) เบราว์เซอร์ส่งให้อัตโนมัติหลังล็อกอิน"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase access token (JWT) สำหรับ API client ที่ไม่มี cookie jar"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "ยังไม่ได้ล็อกอิน หรือ token หมดอายุ/ไม่ถูกต้อง",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "unauthorized",
              "message": "กรุณาเข้าสู่ระบบก่อนเรียกใช้งาน API"
            }
          }
        }
      },
      "BadRequest": {
        "description": "ข้อมูลที่ส่งมาไม่ครบหรือไม่ถูกต้อง",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "ข้อผิดพลาดฝั่งเซิร์ฟเวอร์",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Review": {
        "type": "object",
        "required": [
          "id",
          "userName",
          "rating",
          "comment",
          "date"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "rev-1"
          },
          "userName": {
            "type": "string",
            "example": "คุณสมชาย ใจดี"
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5,
            "example": 5
          },
          "comment": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "example": "2026-08-14"
          }
        }
      },
      "Product": {
        "type": "object",
        "required": [
          "id",
          "name",
          "category",
          "price",
          "stock",
          "description",
          "specs",
          "image",
          "rating",
          "reviewsCount"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "3f6d1c2e-0f1a-4d3b-9a7c-2b8e5d4f1a90"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "originalPrice": {
            "type": "number"
          },
          "stock": {
            "type": "integer"
          },
          "description": {
            "type": "string"
          },
          "specs": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "image": {
            "type": "string",
            "format": "uri"
          },
          "gallery": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "rating": {
            "type": "number"
          },
          "reviewsCount": {
            "type": "integer"
          },
          "badge": {
            "type": "string",
            "enum": [
              "HOT",
              "NEW",
              "SALE",
              "LIMITED"
            ]
          },
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Review"
            }
          },
          "isFeatured": {
            "type": "boolean"
          }
        }
      },
      "ProductInput": {
        "type": "object",
        "required": [
          "name",
          "price",
          "stock"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "originalPrice": {
            "type": "number"
          },
          "stock": {
            "type": "integer"
          },
          "description": {
            "type": "string"
          },
          "specs": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "image": {
            "type": "string",
            "format": "uri"
          },
          "badge": {
            "type": "string",
            "enum": [
              "HOT",
              "NEW",
              "SALE",
              "LIMITED"
            ]
          },
          "isFeatured": {
            "type": "boolean"
          }
        }
      },
      "ProductResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Product"
          }
        }
      },
...
      "StoreSettings": {
        "type": "object",
        "properties": {
          "storeName": {
            "type": "string"
          },
          "isOpen": {
            "type": "boolean"
          },
          "topupReceiverName": {
            "type": "string"
          },
          "topupReceiverAccount": {
            "type": "string"
          },
          "topupBankName": {
            "type": "string"
          },
          "topupMinAmount": {
            "type": "number"
          },
          "topupMaxAmount": {
            "type": "number"
          },
          "topupMaxSlipAgeDays": {
            "type": "integer"
          },
          "freeShippingMin": {
            "type": "number"
          },
          "shippingFee": {
            "type": "number"
          },
          "taxRate": {
            "type": "number"
          }
        }
      }
    }
  }
}